Add files via upload
We were having stability issues, having each instance trying to search for / spin up its own server, so in this version, the syncing-server has been split off as it's own independent program, and the main attendance program is client-only, and can work offline more robustly, as well. There was also a change to the check-in keyboard, where all members appear as selectable tiles, much like searching produce on a self-checkout register in a grocery store.